Move the header back to where it belongs
